[drone_racing] Another fixes to backend of services

+++ services/drone_racing/resources/database/migrations/V04__create_runs.sql
@@ -3,8 +3,8 @@ create table runs
     id          serial primary key,
     program_id  integer references programs (id),
     level_id    integer references levels (id),
-    start_time  integer,
-    finish_time integer,
+    start_time  bigint,
+    finish_time bigint,
     success     boolean,
     score       integer
 )
